generic custom question class created and implemented across existing questions
